spring jdbc template is learned and worked out the source code.
creating a driver manager data source with driverclass name , url , username and password.
creating the jdbc template to which the created data source is added.
the jdbc template created is inturn used in the impl of the DAOs to create, update, retrieve and delete the records in the databases.
by writing a sql and passing a DTO object.

at pom level, following dependencies are added :
spring core, spring context, spring jdbc, mysql connect java